initial work on LogExplorer
